Extract IconButtonWithTooltip for general use.
